fb id to bigint

+++ core/models.py
@@ -2022,7 +2022,7 @@ class UserProfile(models.Model):
     pic_url = models.URLField(blank=True) 
     home_url = models.URLField(blank=True)
     twitter_id = models.CharField(max_length=15, blank=True)
-    facebook_id = models.PositiveIntegerField(null=True)
+    facebook_id = models.BigIntegerField(null=True)
     librarything_id = models.CharField(max_length=31, blank=True)
     badges = models.ManyToManyField('Badge', related_name='holders')
     kindle_email = models.EmailField(max_length=254, blank=True)
